added detail page content
