+++ src/styles/styles.css
@@ -589,8 +589,6 @@ body {
   height: 4rem;
   background-color: var(--main-color);
   border-radius: 30rem 30rem 15rem 15rem;
-  box-shadow: 0.1rem 0.1rem 0.1rem var(--main-color),
-    -0.1rem -0.1rem 0.1rem var(--main-color);
   transition: top 0.4s 0.2s;
 }
